Set the document in a font stack, and read the host's fonts
The text system compiled in one family and could only shape in the faces the app
handed it. It now reads the host's own fonts as well — fontique's platform backend,
which is fontconfig on a Linux desktop — so the document can be set in any family
the machine has, and COOL_SCROLL_FONT takes a whole stack: a comma-separated list
tried in order, the shape a desktop's font configuration is written in, pushed to
the shaper as a source string for fontique to resolve.

Left alone the document is set in the chain a GNOME desktop reaches for — Adwaita
Sans, then Cantarell, then Inter, and so on, ending in the generic that always
resolves — and the readout is set in the stack a GNOME terminal would use rather
than the literal monospace, which used to shape to an empty line.

A stack can resolve to more than one face, since a glyph no family in it can draw
is shaped from whatever can. So a run now carries the face the shaper actually
used, recorded as the layout is converted, and rasterization reads that back.
Before this a fallback glyph would have been rasterized from the requested
family's face, whose glyph ids the run's ids are not — the two halves of a shaped
line have to agree about which face they are talking about.

a_script_the_stack_has_not_got_falls_back_to_its_own_face guards both halves: the
stack resolves, and a script none of its families has got comes back on a face of
its own. The vendored crate's shapes_mixed_weight_runs had asserted the old
contract, that a run's id is the id of the family it was requested with, so it now
asserts the two weights come back as two faces instead.

+++ patches/gpui_parley/examples/raster_cost.rs
@@ -352,7 +352,7 @@ fn main() {
         println!(
             "{name:<11} {:>9} {:>12} {:>12.1?} {:>12.1?}",
             atlas.len(),
-            misses / warm_frames as usize,
+            misses / warm_frames,
             layout_time / warm_frames as u32,
             raster_time / warm_frames as u32,
         );
